+++ packages/framework/src/Component/Domain/Multidomain/MultidomainEntityClassFinderFacade.php
@@ -5,6 +5,7 @@ namespace Shopsys\FrameworkBundle\Component\Domain\Multidomain;
 use Doctrine\ORM\EntityManagerInterface;
 use Shopsys\FrameworkBundle\Component\Entity\EntityNotNullableColumnsFinder;
 use Shopsys\FrameworkBundle\Component\Setting\SettingValue;
+use Shopsys\FrameworkBundle\Model\Category\CategoryDomain;
 use Shopsys\FrameworkBundle\Model\Mail\MailTemplate;
 use Shopsys\FrameworkBundle\Model\Product\Brand\BrandDomain;
 use Shopsys\FrameworkBundle\Model\Product\ProductDomain;
@@ -80,6 +81,7 @@ class MultidomainEntityClassFinderFacade
     {
         return [
             BrandDomain::class,
+            CategoryDomain::class,
             MailTemplate::class,
             ProductDomain::class,
         ];