+++ project-base/src/Shopsys/ShopBundle/Controller/Front/ProductController.php
@@ -171,7 +171,6 @@ class ProductController extends FrontBaseController
             'paginationResult' => $paginationResult,
             'productFilterCountData' => $productFilterCountData,
             'category' => $category,
-            'categoryDomain' => $category->getCategoryDomain($this->domain->getId()),
             'filterForm' => $filterForm->createView(),
             'filterFormSubmited' => $filterForm->isSubmitted(),
             'visibleChildren' => $this->categoryFacade->getAllVisibleChildrenByCategoryAndDomainId($category, $this->domain->getId()),